address review for BypassGovernanceRetention auth
Guard missing bucket_owner, reuse s3_utils for the Bypass header,
evaluate DeleteObjects against bucket and object-wildcard ARNs, trim
docs/tests, and keep a small unit file for Deny-wins and NC edges.
