+++ src/sdk/namespace_fs.js
@@ -2376,8 +2376,7 @@ class NamespaceFS {
     /**
      * check if the object deletion should be blocked by retention lock. if the object is blocked will throw AccessDenied error
      * @param {Object} retention - object retention lock settings
-     * @param {boolean} bypass_governance - authorized Bypass intent from S3 endpoint
-     *   (hosted: IAM/bucket policy; NC: bucket policy). Namespace trusts that decision.
+     * @param {boolean} bypass_governance - authorized Bypass flag from the S3 endpoint
      * @throws {S3Error.AccessDenied} if the object is protected by object lock and the user does not have permission to bypass the lock
      */
     _check_object_retention(fs_context, retention, bypass_governance) {
@@ -2402,7 +2401,7 @@ class NamespaceFS {
      * 2. if the new retention is shorter than the current retention, it cannot be updated and will throw error, unless the user has bypass_governance permission and the current retention mode is GOVERNANCE
      * @param {Object} current_retention - current object retention lock settings
      * @param {Object} new_retention - new object retention lock settings
-     * @param {boolean} bypass_governance - authorized Bypass intent from S3 endpoint
+     * @param {boolean} bypass_governance - authorized Bypass flag from the S3 endpoint
      * @throws {S3Error.AccessDenied} if the object is protected by object lock and the user does not have permission to bypass the lock
      */
     _compare_object_retention(fs_context, current_retention, new_retention, bypass_governance) {
