+++ src/test/integration_tests/api/s3/test_s3_worm.js
@@ -1054,7 +1054,6 @@ mocha.describe('s3 worm', function() {
     });
 
     // NC bucket owner may Bypass without a bucket-policy grant (AWS account-root analog).
-    // Non-owner NC accounts still need bucket-policy Bypass (see unit coverage).
     mocha.describe('NC - bucket owner Bypass without bucket policy', function() {
         let version_id;
         mocha.before(async function() {
