+++ src/test/unit_tests/internal/test_object_server.test.js
@@ -543,13 +543,3 @@ describe('object_server - update_bulk_delete_results', () => {
         expect(results[3]).toHaveProperty('seq', 103);
     });
 });
-
-describe('object_server._can_bypass_governance', () => {
-    const { _can_bypass_governance } = object_server.__testing;
-
-    it('returns the RPC bypass_governance flag', () => {
-        expect(_can_bypass_governance({ rpc_params: { bypass_governance: true } })).toBe(true);
-        expect(_can_bypass_governance({ rpc_params: { bypass_governance: false } })).toBe(false);
-        expect(_can_bypass_governance({ rpc_params: {} })).toBe(false);
-    });
-});
