+++ src/test/unit_tests/util_functions_tests/test_access_policy_utils.test.js
@@ -90,23 +90,6 @@ describe('access_policy_utils', () => {
                 ).resolves.toBeUndefined();
             });
 
-            it('should accept s3:BypassGovernanceRetention', async () => {
-                const policy = make_policy({
-                    action: [
-                        's3:PutObjectRetention',
-                        's3:DeleteObject',
-                        's3:BypassGovernanceRetention',
-                    ],
-                    resource: [
-                        `arn:aws:s3:::${BUCKET_NAME}`,
-                        `arn:aws:s3:::${BUCKET_NAME}/*`,
-                    ],
-                });
-                await expect(
-                    access_policy_utils.validate_bucket_policy(policy, BUCKET_NAME, account_handler_allow_all)
-                ).resolves.toBeUndefined();
-            });
-
             it('should accept a Deny statement with NotPrincipal', async () => {
                 const policy = make_policy({
                     effect: 'Deny',
